dom/tests/mochitest/cowl/test_privileges.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <p id="display"></p>
  <pre id="test">

  <script class="testbody" type="application/javascript">
  SimpleTest.waitForExplicitFinish();

  var completedTests = 0;
  var passedTests = 0;

  function ok_wrapper(result, desc) {
    ok(result, desc);

    completedTests++;

    if (result) {
      passedTests++;
    }

    if (completedTests == 6) {
      is(passedTests, completedTests, "There are " + completedTests + " privilege tests that should pass");

      SimpleTest.finish();
    }
  }

  function receiveMessage(event)
  {
    ok_wrapper(event.data.ok, event.data.desc);
  }

  window.addEventListener("message", receiveMessage, false);

  </script>
  <iframe src="file_privileges.html"></iframe>
</pre>
</body>
</html>
